Speed up and unify flash programming
This patch creates a FlashBuilder class which can be used to build a
list of data blocks to program.  It also allows programming to occur
faster since FlashBuider automatically determines the best way to
program the data passed to it.  When performing page erase programming,
pages with identical data do not get programmed.  When a chip erase
is preformed, pages filled with 0xFF do not get programmed.

This patch also takes the flashing code which gdbserver uses and
replaces it with a call to flash's flashBlock function, which increases
performance and unifies the flashing code.

Finally, this patch adds flash_test.py to test the different flashing
modes.